Fix createHiddenInputs [#111713764]
Right now we're not creating hidden fields when the select field has no
value, so it's hard to check if the user is trying to delete all
entries from the specific select field. It'll be better to create an
empty hidden field so we can receive `['']` in the backend and treat it
accordingly (deleting all values from the field).